Prepare to refactor CSS with initial goal of making it easier to understand and have font-sizes which aren't way too big.
The current CSS from casper-two is a total disaster.
Unminified, it is 1864 lines and has 30 uses of
"@media (max-width:123px) {...}" to make various size
and other options work with different browser window
sizes.  I guess this must've come from a tool because
it would take a sick human to produce this trash.

This commit has some convenience copies in
static/css/.

We are also removing a dependency on
cloudflare for a small file that seems to be related
to syntax highlighting.